Modification type of LoginController return for error messages.

+++ src/main/java/mailjet/api/ApiCampaign.java
@@ -13,7 +13,7 @@ public class ApiCampaign {
 
     public Date SendStartAt;
     
-    public Integer SpamassScore;
+    public float SpamassScore;
 
     public ApiCampaign() {
     }
